ai-rag-service/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
chromadb
sentence-transformers
